Restructure Flask app into a package
Warning: Big commit

I've restructured the Flask application to follow the guidelines on
Flask's site: http://flask.pocoo.org/docs/0.10/patterns/packages/.
In summary, they suggest to use a package instead of a module, and to
move route definitions outside of the Flask application initialization.
This ends up being a lot cleaner.

Major changes:
- Move all files into a folder, so that we can import the entire site as
  a module.
- Make flask application a package, and move app initialization into
  `__init__.py`. This means we can have our route definitions exist
  separate from our app initialization. Yay!
- We technically have circular imports, but Flask says that's okay
  (because we're not using the views; only importing them at the bottom)
- To load the site for testing, run `python run_server.py` in the
  repository root directory. This automatically enables debugging and
  uses a different port. We can do this because this isn't how wsgi
  loads the site, which means we can later customize `run_server.py` to
  do whatever makes testing easy for us!

Note: Moving git submodules isn't easy with our version of Git, so I had
to do it manually. I followed instructions at
https://stackoverflow.com/questions/4604486/how-do-i-move-an-existing-git-submodule-within-a-git-repository.
