+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "static/constitution"]
-	path = static/constitution
-	url = https://github.com/RuddockHouse/RuddockConstitution.git
+[submodule "RuddockWebsite/static/constitution"]
+	path = RuddockWebsite/static/constitution
+	url = https://github.com/RuddockHouse/RuddockConstitution
