+++ RuddockWebsite/common_helpers.py
