+++ RuddockWebsite/modules/admin/__init__.py
@@ -1,4 +1,4 @@
 from flask import Blueprint
 blueprint = Blueprint('admin', __name__, template_folder='templates')
 
-import modules.admin.routes
+import RuddockWebsite.modules.admin.routes
