+++ RuddockWebsite/modules/hassle/__init__.py
@@ -1,4 +1,4 @@
 from flask import Blueprint
 blueprint = Blueprint('hassle', __name__, template_folder='templates')
 
-import modules.hassle.routes
+import RuddockWebsite.modules.hassle.routes
