+++ RuddockWebsite/modules/hassle/templates/hassle_new_participants.html
