+++ RuddockWebsite/modules/users/__init__.py
@@ -1,4 +1,4 @@
 from flask import Blueprint
 blueprint = Blueprint('users', __name__, template_folder='templates')
 
-import modules.users.routes
+import RuddockWebsite.modules.users.routes
